+++ src/main/java/org/kumoricon/registration/staff/CheckInController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import java.io.IOException;
 import java.time.Instant;
+import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -29,14 +30,33 @@ public class CheckInController {
 
     @RequestMapping(value="/staff")
     @PreAuthorize("hasAuthority('staff_check_in')")
-    public String staffList(Model model) {
+    public String staffList(Model model,
+                            @RequestParam(required = false, defaultValue = "false") boolean checkedIn ) {
         List<Staff> staff = staffRepository.findAll();
-        int checkedIn = 0;          // Just count them here to save another database call, since
-        for (Staff s : staff) {     // We need the whole list anyway
-            if (s.getCheckedIn()) checkedIn++;
+        // TODO: filtering this in code instead of the database is terrible and needs to be
+        // refactored, but this is the quickest way to do what I need to do right now.
+        // The whole thing should be changed to do the search server-side anyway (though maybe with some
+        // search-as-you-type callbacks?) because some of the clients are ridiculously old and slow
+        // Seriously, my phone has a faster CPU.
+        List<Staff> outputCheckedIn = new ArrayList<>();
+        List<Staff> outputNotCheckedIn = new ArrayList<>();
+        int checkedInCount = 0;          // Just count them here to save another database call
+        for (Staff s : staff) {
+            if (s.getCheckedIn()) {
+                checkedInCount++;
+                outputCheckedIn.add(s);
+            } else {
+                outputNotCheckedIn.add(s);
+            }
         }
-        model.addAttribute("staff", staff);
+        if (checkedIn) {
+            model.addAttribute("staff", outputCheckedIn);
+        } else {
+            model.addAttribute("staff", outputNotCheckedIn);
+        }
         model.addAttribute("checkedIn", checkedIn);
+        model.addAttribute("totalCount", staff.size());
+        model.addAttribute("checkedInCount", checkedInCount);
 
         return "staff/stafflist";
     }